Rename folder `table-pivot` by `pivot-table`
