+++ projects/ontimize-web-ngx/src/lib/components/index.ts
@@ -40,4 +40,4 @@ export * from './expandable-container/o-expandable-container.component';
 export * from './expandable-container/o-expandable-container.module';
 export * from './tree/index';
 export * from './o-skeleton.component';
-export * from './table-pivot/index';
+export * from './pivot-table/index';
