+++ projects/ontimize-web-ngx/src/lib/components/pivot-table/pivot-table.component.scss
