+++ projects/ontimize-web-ngx/src/lib/components/pivot-table/select-function-dialog/select-function-dialog.component.html
