+++ projects/ontimize-web-ngx/src/lib/components/pivot-table/table-pivot.theme.scss
